requirements.txt
Django==1.9
django-rest-framework-mongoengine==3.3.0
djangorestframework==3.3.1
mongoengine==0.10.6
MySQL-python==1.2.5
phonenumbers==7.2.5
pycountry==1.20
pymongo==3.2.1
wheel==0.24.0
